fix(212): Fix Jitpack dependencies for kamelet-catalog-1.6
Kamelets in this branch `kamelet-catalog-1.6` must not reference Jitpack dependencies with `master-SNAPSHOT` but need to use `kamelet-catalog-1.6-SNAPSHOT`. This ensures that the Kamelets only pull utility classes from this branch via Jitpack. The `master` branch may then move on and introduce breaking changes in utility classes in the future without breaking Kamelets in this branch.